Add script execution of ON_NEW_DOCUMENT phase
Refactor script execution

Scripts are executed in two phases. First ON_NEW_DOCUMENT and then
ON_LOAD. The only difference between the two phases is that in the
ON_NEW_DOCUMENT phase a new execution context is created before
execution of every script while in the ON_LOAD phase the same
execution context is reused.

The execution contexts are now created in isolation so we can be
sure they are sane and that names do not collide with names in the
page.

The script execution logic is moved to a new package: script.