feat: add support for async browser scripts
This commit makes it possible to define browser scripts of type on_load
and on_new_document as async functions.

Scripts that are prefixed with 'async ' will now be handled as returning
a promise.

The motivation for this change is first and firemost the ability to call
other async functions, eg. use the Fetch API or set timeouts (sleep).