feat: enable async extract outlinks scripts
This commit makes it possible to define
scripts of type extract_outlinks as an  async
Immediatly-invoked Function Expression (IIFE).

Scripts that have the '(async ' prefix is treated as returning a
promise. For example is it now possible to run a script as follows:

(async funcction() {
  // fetch sitemap.xml
  const response = await fetch("http://example.com/movies.json");
  return parseSiteMapToLinks(await response.json());
})()

This change makes it possible to use async functions when extracting
outlinks, e.g. fetch and parse sitemaps.